SCHED-99: Remove old modules and add missing vskyutils functions for Brightness
